[TIMOB-25826] Fix test and fixture files
