+++ tests/Resources/fixtures/empty-double.js
@@ -1,1 +1,2 @@
-module.exports = "";
+'use strict';
+module.exports = '';
