+++ tests/Resources/fixtures/empty-single.js
@@ -1,1 +1,2 @@
+'use strict';
 module.exports = '';
