+++ tests/Resources/fixtures/false.js
@@ -1,1 +1,2 @@
+'use strict';
 module.exports = false;
