+++ tests/Resources/fixtures/null.js
@@ -1,1 +1,2 @@
+'use strict';
 module.exports = null;
