+++ tests/Resources/fixtures/undefined.js
@@ -1,1 +1,2 @@
+'use strict';
 module.exports = undefined;
